Make DllImport target function local.
