+++ DllImportGenerator/DllImportGenerator/StubCodeGenerator.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-
+using System.Runtime.InteropServices;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -115,7 +115,7 @@ namespace Microsoft.Interop
             }
         }
 
-        public (BlockSyntax Code, MethodDeclarationSyntax DllImport) GenerateSyntax()
+        public BlockSyntax GenerateSyntax()
         {
             string dllImportName = stubMethod.Name + "__PInvoke__";
             var setupStatements = new List<StatementSyntax>();
@@ -340,20 +340,22 @@ namespace Microsoft.Interop
             var codeBlock = Block(UnsafeStatement(Block(allStatements)));
 
             // Define P/Invoke declaration
-            var dllImport = MethodDeclaration(retMarshaller.Generator.AsNativeType(retMarshaller.TypeInfo), dllImportName)
+            var dllImport = LocalFunctionStatement(retMarshaller.Generator.AsNativeType(retMarshaller.TypeInfo), dllImportName)
                 .AddModifiers(
                     Token(SyntaxKind.ExternKeyword),
-                    Token(SyntaxKind.PrivateKeyword),
                     Token(SyntaxKind.StaticKeyword),
                     Token(SyntaxKind.UnsafeKeyword))
-                .WithSemicolonToken(Token(SyntaxKind.SemicolonToken));
+                .WithSemicolonToken(Token(SyntaxKind.SemicolonToken))
+                .WithAttributeLists(
+                    SingletonList(AttributeList(
+                        SingletonSeparatedList(CreateDllImportAttributeForTarget(GetTargetDllImportDataFromStubData())))));
             foreach (var marshaller in paramMarshallers)
             {
                 ParameterSyntax paramSyntax = marshaller.Generator.AsParameter(marshaller.TypeInfo);
                 dllImport = dllImport.AddParameterListParameters(paramSyntax);
             }
 
-            return (codeBlock, dllImport);
+            return codeBlock.AddStatements(dllImport);
 
             List<StatementSyntax> GetStatements(Stage stage)
             {
@@ -398,6 +400,132 @@ namespace Microsoft.Interop
                     generator.AsNativeType(info),
                     native));
             }
+        }
+
+        private static AttributeSyntax CreateDllImportAttributeForTarget(DllImportStub.GeneratedDllImportData targetDllImportData)
+        {
+            var newAttributeArgs = new List<AttributeArgumentSyntax>
+            {
+                AttributeArgument(LiteralExpression(
+                    SyntaxKind.StringLiteralExpression,
+                    Literal(targetDllImportData.ModuleName))),
+                AttributeArgument(
+                    NameEquals(nameof(DllImportAttribute.EntryPoint)),
+                    null,
+                    CreateStringExpressionSyntax(targetDllImportData.EntryPoint))
+            };
+
+            if (targetDllImportData.IsUserDefined.HasFlag(DllImportStub.DllImportMember.BestFitMapping))
+            {
+                var name = NameEquals(nameof(DllImportAttribute.BestFitMapping));
+                var value = CreateBoolExpressionSyntax(targetDllImportData.BestFitMapping);
+                newAttributeArgs.Add(AttributeArgument(name, null, value));
+            }
+            if (targetDllImportData.IsUserDefined.HasFlag(DllImportStub.DllImportMember.CallingConvention))
+            {
+                var name = NameEquals(nameof(DllImportAttribute.CallingConvention));
+                var value = CreateEnumExpressionSyntax(targetDllImportData.CallingConvention);
+                newAttributeArgs.Add(AttributeArgument(name, null, value));
+            }
+            if (targetDllImportData.IsUserDefined.HasFlag(DllImportStub.DllImportMember.CharSet))
+            {
+                var name = NameEquals(nameof(DllImportAttribute.CharSet));
+                var value = CreateEnumExpressionSyntax(targetDllImportData.CharSet);
+                newAttributeArgs.Add(AttributeArgument(name, null, value));
+            }
+            if (targetDllImportData.IsUserDefined.HasFlag(DllImportStub.DllImportMember.ExactSpelling))
+            {
+                var name = NameEquals(nameof(DllImportAttribute.ExactSpelling));
+                var value = CreateBoolExpressionSyntax(targetDllImportData.ExactSpelling);
+                newAttributeArgs.Add(AttributeArgument(name, null, value));
+            }
+            if (targetDllImportData.IsUserDefined.HasFlag(DllImportStub.DllImportMember.PreserveSig))
+            {
+                var name = NameEquals(nameof(DllImportAttribute.PreserveSig));
+                var value = CreateBoolExpressionSyntax(targetDllImportData.PreserveSig);
+                newAttributeArgs.Add(AttributeArgument(name, null, value));
+            }
+            if (targetDllImportData.IsUserDefined.HasFlag(DllImportStub.DllImportMember.SetLastError))
+            {
+                var name = NameEquals(nameof(DllImportAttribute.SetLastError));
+                var value = CreateBoolExpressionSyntax(targetDllImportData.SetLastError);
+                newAttributeArgs.Add(AttributeArgument(name, null, value));
+            }
+            if (targetDllImportData.IsUserDefined.HasFlag(DllImportStub.DllImportMember.ThrowOnUnmappableChar))
+            {
+                var name = NameEquals(nameof(DllImportAttribute.ThrowOnUnmappableChar));
+                var value = CreateBoolExpressionSyntax(targetDllImportData.ThrowOnUnmappableChar);
+                newAttributeArgs.Add(AttributeArgument(name, null, value));
+            }
+
+            // Create new attribute
+            return Attribute(
+                ParseName(typeof(DllImportAttribute).FullName),
+                AttributeArgumentList(SeparatedList(newAttributeArgs)));
+
+            static ExpressionSyntax CreateBoolExpressionSyntax(bool trueOrFalse)
+            {
+                return LiteralExpression(
+                    trueOrFalse
+                        ? SyntaxKind.TrueLiteralExpression
+                        : SyntaxKind.FalseLiteralExpression);
+            }
+
+            static ExpressionSyntax CreateStringExpressionSyntax(string str)
+            {
+                return LiteralExpression(
+                    SyntaxKind.StringLiteralExpression,
+                    Literal(str));
+            }
+
+            static ExpressionSyntax CreateEnumExpressionSyntax<T>(T value) where T : Enum
+            {
+                return MemberAccessExpression(
+                    SyntaxKind.SimpleMemberAccessExpression,
+                    IdentifierName(typeof(T).FullName),
+                    IdentifierName(value.ToString()));
+            }
+        }
+
+        DllImportStub.GeneratedDllImportData GetTargetDllImportDataFromStubData()
+        {
+            DllImportStub.DllImportMember membersToForward = DllImportStub.DllImportMember.All
+                               // https://docs.microsoft.com/dotnet/api/system.runtime.interopservices.dllimportattribute.preservesig
+                               // If PreserveSig=false (default is true), the P/Invoke stub checks/converts a returned HRESULT to an exception.
+                               & ~DllImportStub.DllImportMember.PreserveSig
+                               // https://docs.microsoft.com/dotnet/api/system.runtime.interopservices.dllimportattribute.setlasterror
+                               // If SetLastError=true (default is false), the P/Invoke stub gets/caches the last error after invoking the native function.
+                               & ~DllImportStub.DllImportMember.SetLastError;
+            if (options.GenerateForwarders())
+            {
+                membersToForward = DllImportStub.DllImportMember.All;
+            }
+
+            var targetDllImportData = new DllImportStub.GeneratedDllImportData
+            {
+                CharSet = dllImportData.CharSet,
+                BestFitMapping = dllImportData.BestFitMapping,
+                CallingConvention = dllImportData.CallingConvention,
+                EntryPoint = dllImportData.EntryPoint,
+                ModuleName = dllImportData.ModuleName,
+                ExactSpelling = dllImportData.ExactSpelling,
+                SetLastError = dllImportData.SetLastError,
+                PreserveSig = dllImportData.PreserveSig,
+                ThrowOnUnmappableChar = dllImportData.ThrowOnUnmappableChar,
+                IsUserDefined = dllImportData.IsUserDefined & membersToForward
+            };
+
+            // If the EntryPoint property is not set, we will compute and
+            // add it based on existing semantics (i.e. method name).
+            //
+            // N.B. The export discovery logic is identical regardless of where
+            // the name is defined (i.e. method name vs EntryPoint property).
+            if (!targetDllImportData.IsUserDefined.HasFlag(DllImportStub.DllImportMember.EntryPoint))
+            {
+                targetDllImportData.EntryPoint = stubMethod.Name;
+            }
+
+            return targetDllImportData;
         }
     }
 }
